Update to add clang support
Now the project can be successfully compiled on iOS's devtools.

+++ ProceduralCityGen/work/src/aStarSearchAlgo.cpp
@@ -109,7 +109,7 @@ void AStarSearchAlgo::aStarSearch(roadNode *start, roadNode *goal) {
 
 					// Add a new fringeNode to the queue
 					fringeNode add = { &m_allNodes[neighID],
-							&m_allNodes[nodeID], costToNeigh, estTotal };
+							&m_allNodes[nodeID], costToNeigh, (float)estTotal };
 					fringe.push(add);
 				}
 			}
